JBIDE-14591 https://issues.jboss.org/browse/JBIDE-14591
Implemented categories for Properties view for html elements.
