+++ plugins/org.jboss.tools.jst.jsp/src/org/jboss/tools/jst/jsp/outline/JSPPropertySourceAdapter.java
@@ -28,6 +28,7 @@ import org.eclipse.ui.views.properties.IPropertyDescriptor;
 import org.eclipse.ui.views.properties.IPropertySheetEntry;
 import org.eclipse.ui.views.properties.IPropertySource;
 import org.eclipse.ui.views.properties.IPropertySource2;
+import org.eclipse.ui.views.properties.PropertyDescriptor;
 import org.eclipse.ui.views.properties.TextPropertyDescriptor;
 import org.eclipse.wst.sse.core.internal.provisional.INodeAdapter;
 import org.eclipse.wst.sse.core.internal.provisional.INodeNotifier;
@@ -44,13 +45,16 @@ import org.eclipse.wst.xml.core.internal.provisional.document.IDOMNode;
 import org.eclipse.wst.xml.ui.internal.XMLUIMessages;
 import org.eclipse.wst.xml.ui.internal.properties.EnumeratedStringPropertyDescriptor;
 import org.jboss.tools.common.model.ui.ModelUIPlugin;
+import org.jboss.tools.common.util.FileUtil;
 import org.jboss.tools.jst.jsp.JspEditorPlugin;
 import org.jboss.tools.jst.jsp.contentassist.computers.FaceletsELCompletionProposalComputer;
 import org.jboss.tools.jst.jsp.contentassist.computers.JspELCompletionProposalComputer;
 import org.jboss.tools.jst.jsp.editor.IVisualController;
 import org.jboss.tools.jst.jsp.jspeditor.JSPMultiPageEditor;
+import org.jboss.tools.jst.web.html.HTMLConstants;
 import org.jboss.tools.jst.web.kb.IPageContext;
 import org.jboss.tools.jst.web.kb.KbQuery;
+import org.jboss.tools.jst.web.kb.KbQuery.Tag;
 import org.jboss.tools.jst.web.kb.KbQuery.Type;
 import org.jboss.tools.jst.web.kb.PageProcessor;
 import org.jboss.tools.jst.web.kb.taglib.IAttribute;
@@ -98,14 +102,17 @@ public class JSPPropertySourceAdapter implements INodeAdapter, IPropertySource, 
 	JspELCompletionProposalComputer processor;
 	int offset = 0;
 	KbQuery kbQuery, kbQueryAttr;
-	private Set attributeNames = new HashSet();
+	private Set<String> attributeNames = new HashSet<String>();
+
+		boolean isHTML5 = false;
+		ICategoryProvider categoryProvider = null;
 
 	public JSPPropertySourceAdapter(INodeNotifier target) {
 		setTarget(target);
 	}
 	
-	Map getWeights() {
-		return sorter == null ? new HashMap() : sorter.weights;
+	Map<String,Integer> getWeights() {
+		return sorter == null ? new HashMap<String,Integer>() : sorter.weights;
 	}
 	
 	public void setTarget(INodeNotifier target) {
@@ -123,6 +130,8 @@ public class JSPPropertySourceAdapter implements INodeAdapter, IPropertySource, 
 			offset = ((IDOMNode)fNode).getStartOffset() + ("" + fNode.getNodeName()).length(); //$NON-NLS-1$
 		}
 
+		isHTML5 = false;
+		categoryProvider = null;
 		if(fNode instanceof Node) {
 			ITextViewer viewer = getTextViewer();
 			// Jeremy: JBIDE-9949: This prevents invocation of CA Proposals Computation in case 
@@ -139,7 +148,14 @@ public class JSPPropertySourceAdapter implements INodeAdapter, IPropertySource, 
 				kbQuery.setMask(true); 
 				kbQueryAttr = createKbQuery(processor);
 			}
+			
+			isHTML5 = FileUtil.isDoctypeHTML(viewer.getDocument().get());
+			if(isHTML5) {
+				categoryProvider = new JQueryCategoryProvider();
+				categoryProvider.init(kbQuery);
+			}
 		}
+		
 	}
 	
 	//TODO move to helper
@@ -182,9 +198,15 @@ public class JSPPropertySourceAdapter implements INodeAdapter, IPropertySource, 
 		} else if (fDescriptors == null || fDescriptors.length == 0) {
 			attributeNamesChanged();
 			fDescriptors = createPropertyDescriptors();
+			if(categoryProvider != null && sorter != null) {
+				categoryProvider.fillAttributeWeights(getWeights());
+			}
 		} else {
 			if(attributeNamesChanged()) {
 				updatePropertyDescriptors();
+				if(categoryProvider != null && sorter != null) {
+					categoryProvider.fillAttributeWeights(getWeights());
+				}
 			}
 //			return fDescriptors;
 		}
@@ -319,40 +341,49 @@ public class JSPPropertySourceAdapter implements INodeAdapter, IPropertySource, 
 	protected KbQuery createKbQuery(JspELCompletionProposalComputer processor) {
 		KbQuery kbQuery = new KbQuery();
 
-		String[] parentTags = processor.getParentTags(false);
-		parentTags = add(parentTags, fNode.getNodeName());
+		KbQuery.Tag[] parentTags = processor.getParentTagsWithAttributes(true);
 		kbQuery.setPrefix(getPrefix());
 		kbQuery.setUri(processor.getUri(getPrefix()));
-		kbQuery.setParentTags(parentTags);
+		kbQuery.setParentTagsWithAttributes(parentTags);
 		kbQuery.setParent(fNode.getNodeName());
 		kbQuery.setMask(false); 
 		kbQuery.setType(Type.ATTRIBUTE_NAME);
 		kbQuery.setOffset(offset);
 		kbQuery.setValue("");  //$NON-NLS-1$
 		kbQuery.setStringQuery(""); //$NON-NLS-1$
+
+		Tag[] tags = kbQuery.getParentTagsWithAttributes();
+		if(tags.length > 0) {
+			Tag current = tags[tags.length - 1];
+			kbQuery.setAttributes(current.getAttributes());
+		}
 		
 		return kbQuery;
 	}
 
-	private String[] add(String[] result, String v) {
-		String[] result1 = new String[result.length + 1];
-		System.arraycopy(result, 0, result1, 0, result.length);
-		result1[result.length] = v;
-		return result1;
-	}
 	private String getPrefix() {
 		int i = fNode.getNodeName().indexOf(':');
 		return i < 0 ? null : fNode.getNodeName().substring(0, i);
 	}
 
-
-	private String getCategory(CMAttributeDeclaration attrDecl) {
+	private String getCategory(String attrName, CMAttributeDeclaration attrDecl) {
+		String result = getCategoryByName(attrName);
+		if(result != null) {
+			return result;
+		}
 		if (attrDecl != null) {
 			if (attrDecl.supports("category")) { //$NON-NLS-1$
 				return (String) attrDecl.getProperty("category"); //$NON-NLS-1$
 			}
 		}
 		return CATEGORY_ATTRIBUTES;
+	}
+
+	private String getCategoryByName(String attrName) {
+		if(categoryProvider != null && fNode != null && fNode.getNodeType() == Node.ELEMENT_NODE) {
+			return categoryProvider.getCategory(attrName);
+		}
+		return null;
 	}
 
 	private CMElementDeclaration getDeclaration() {
@@ -367,7 +398,7 @@ public class JSPPropertySourceAdapter implements INodeAdapter, IPropertySource, 
 	}
 
 	private IPropertyDescriptor createJSPPropertyDescriptor(IAttribute d, String attributeName, boolean hideOnFilter) {
-		if(d != null && sorter != null) {
+		if(d != null && sorter != null && categoryProvider == null) {
 			if(d.isRequired()) sorter.setWeight(attributeName, 2);
 			else if(d.isPreferable()) sorter.setWeight(attributeName, 1);
 			else sorter.setWeight(attributeName, 0);
@@ -381,16 +412,29 @@ public class JSPPropertySourceAdapter implements INodeAdapter, IPropertySource, 
 		context.put("processor", processor); //$NON-NLS-1$
 		context.put("queryFactory", queryFactory); //$NON-NLS-1$
 		JSPPropertyDescriptor descriptor = new JSPPropertyDescriptor(context, attributeName, attributeName);
-		descriptor.setCategory(getCategory(null));
+		
+		setCategory(descriptor, attributeName, null);
+		
 		descriptor.setDescription(attributeName);
 		return descriptor;
+	}
+
+	
+	private void setCategory(PropertyDescriptor descriptor, String attributeName, CMAttributeDeclaration attrDecl) {
+		String category = getCategory(attributeName, attrDecl);
+		descriptor.setCategory(category);
+		if(SET_EXPERT_FILTER && categoryProvider != null && categoryProvider.isExpert(category)) {
+			descriptor.setFilterFlags(new String[]{IPropertySheetEntry.FILTER_ID_EXPERT});
+		}
 	}
 
 	private IPropertyDescriptor createDefaultPropertyDescriptor(String attributeName, boolean hideOnFilter) {
 		// The descriptor class used here is also used in
 		// updatePropertyDescriptors()
 		TextPropertyDescriptor descriptor = new TextPropertyDescriptor(attributeName, attributeName);
-		descriptor.setCategory(getCategory(null));
+
+		setCategory(descriptor, attributeName, null);
+
 		descriptor.setDescription(attributeName);
 //		if (hideOnFilter && SET_EXPERT_FILTER)
 //			descriptor.setFilterFlags(new String[]{IPropertySheetEntry.FILTER_ID_EXPERT});
@@ -817,33 +861,36 @@ public class JSPPropertySourceAdapter implements INodeAdapter, IPropertySource, 
 	}
 
 	private IPropertyDescriptor createTextPropertyDescriptor(CMAttributeDeclaration attrDecl) {
-		return createTextPropertyDescriptor(attrDecl.getAttrName(), getCategory(attrDecl), attrDecl.getUsage());
-	}
+		String name = attrDecl.getAttrName();
 
-	private IPropertyDescriptor createTextPropertyDescriptor(String name, String category, int usage) {
 		TextPropertyDescriptor descriptor = new TextPropertyDescriptor(name, name);
-		descriptor.setCategory(category);
+		
+		setCategory(descriptor, name, attrDecl);
+//		int usage = attrDecl.getUsage();
+//		if (usage != CMAttributeDeclaration.REQUIRED && SET_EXPERT_FILTER)
+//			descriptor.setFilterFlags(new String[]{IPropertySheetEntry.FILTER_ID_EXPERT});
 		descriptor.setDescription(name);
-		if (usage != CMAttributeDeclaration.REQUIRED && SET_EXPERT_FILTER)
-			descriptor.setFilterFlags(new String[]{IPropertySheetEntry.FILTER_ID_EXPERT});
 		return descriptor;
 	}
 
 	private IPropertyDescriptor createEnumeratedPropertyDescriptor(CMAttributeDeclaration attrDecl, CMDataType valuesHelper) {
 		// the displayName MUST be set
-	    	EnumeratedStringPropertyDescriptor descriptor = new EnumeratedStringPropertyDescriptor(attrDecl.getAttrName(), attrDecl.getAttrName(), _getValidStrings(attrDecl, valuesHelper));
-		descriptor.setCategory(getCategory(attrDecl));
-		descriptor.setDescription(attrDecl.getAttrName());
-		if (attrDecl.getUsage() != CMAttributeDeclaration.REQUIRED && SET_EXPERT_FILTER)
-			descriptor.setFilterFlags(new String[]{IPropertySheetEntry.FILTER_ID_EXPERT});
+	    EnumeratedStringPropertyDescriptor descriptor = new EnumeratedStringPropertyDescriptor(attrDecl.getAttrName(), attrDecl.getAttrName(), _getValidStrings(attrDecl, valuesHelper));
+	    String name = attrDecl.getAttrName();
+	    setCategory(descriptor, name, attrDecl);
+//	    int usage = attrDecl.getUsage();
+//		if (attrDecl.getUsage() != CMAttributeDeclaration.REQUIRED && SET_EXPERT_FILTER)
+//			descriptor.setFilterFlags(new String[]{IPropertySheetEntry.FILTER_ID_EXPERT});
+		descriptor.setDescription(name);
 		return descriptor;
 	}
 
 	private IPropertyDescriptor createFixedPropertyDescriptor(CMAttributeDeclaration attrDecl, CMDataType helper) {
 		// the displayName MUST be set
 		EnumeratedStringPropertyDescriptor descriptor = new EnumeratedStringPropertyDescriptor(attrDecl.getNodeName(), attrDecl.getNodeName(), _getValidFixedStrings(attrDecl, helper));
-		descriptor.setCategory(getCategory(attrDecl));
-		descriptor.setDescription(attrDecl.getAttrName());
+		String name = attrDecl.getAttrName();
+		setCategory(descriptor, name, attrDecl);
+		descriptor.setDescription(name);
 		return descriptor;
 	}
 
@@ -863,7 +910,7 @@ public class JSPPropertySourceAdapter implements INodeAdapter, IPropertySource, 
 
 	private String[] _getValidStrings(CMAttributeDeclaration attrDecl, CMDataType valuesHelper) {
 		String attributeName = attrDecl.getAttrName();
-		List values = new ArrayList(1);
+		List<String> values = new ArrayList<String>(1);
 		boolean currentValueKnown = false;
 		boolean checkIfCurrentValueIsKnown = (fNode.getAttributes() != null && fNode.getAttributes().getNamedItem(attributeName) != null && fNode.getAttributes().getNamedItem(attributeName).getNodeValue() != null);
 		String currentValue = null;
@@ -903,13 +950,13 @@ public class JSPPropertySourceAdapter implements INodeAdapter, IPropertySource, 
 		if (checkIfCurrentValueIsKnown && !currentValueKnown && currentValue != null && currentValue.length() > 0)
 			values.add(currentValue);
 		String[] validStrings = new String[values.size()];
-		validStrings = (String[]) values.toArray(validStrings);
+		validStrings = values.toArray(validStrings);
 		return validStrings;
 	}
 
 	private String[] _getValidFixedStrings(CMAttributeDeclaration attrDecl, CMDataType helper) {
 		String attributeName = attrDecl.getAttrName();
-		List values = new ArrayList(1);
+		List<String> values = new ArrayList<String>(1);
 		String impliedValue = helper.getImpliedValue();
 		if (impliedValue != null)
 			values.add(impliedValue);
