+++ plugins/org.jboss.tools.jst.web.ui/src/org/jboss/tools/jst/web/ui/palette/html/jquery/wizard/JQueryConstants.java
@@ -10,157 +10,14 @@
  ******************************************************************************/ 
 package org.jboss.tools.jst.web.ui.palette.html.jquery.wizard;
 
-import org.jboss.tools.jst.web.ui.palette.html.wizard.HTMLConstants;
+import org.jboss.tools.jst.web.html.JQueryHTMLConstants;
 
 /**
  * 
  * @author Viacheslav Kabanovich
  *
  */
-public interface JQueryConstants extends HTMLConstants {
-	public String ATTR_DATA_AUTODIVIDERS = "data-autodividers";
-	public String ATTR_DATA_CLEAR_BTN = "data-clear-btn";
-	public String ATTR_DATA_CLOSE_BTN = "data-close-btn";
-	public String ATTR_DATA_COLLAPSED = "data-collapsed";
-	public String ATTR_DATA_COLLAPSED_ICON = "data-collapsed-icon";
-	public String ATTR_DATA_CONTENT_THEME = "data-content-theme";
-	public String ATTR_DATA_DISPLAY = "data-display";
-	public String ATTR_DATA_EXPANDED_ICON = "data-expanded-icon";
-	public String ATTR_DATA_FILTER = "data-filter";
-	public String ATTR_DATA_FULL_SCREEN = "data-fullscreen";
-	public String ATTR_DATA_HIGHLIGHT = "data-highlight";
-	public String ATTR_DATA_ICON = "data-icon";
-	public String ATTR_DATA_ICONPOS = "data-iconpos";
-	public String ATTR_DATA_INLINE = "data-inline";
-	public String ATTR_DATA_INSET = "data-inset";
-	public String ATTR_DATA_MINI = "data-mini";
-	public String ATTR_DATA_MODE = "data-mode";
-	public String ATTR_DATA_OVERLAY_THEME = "data-overlay-theme";
-	public String ATTR_DATA_POSITION = "data-position";
-	public String ATTR_DATA_POSITION_FIXED = "data-position-fixed";
-	public String ATTR_DATA_PRIORITY = "data-priority";
-	public String ATTR_DATA_REL = "data-rel";
-	public String ATTR_DATA_ROLE = "data-role";
-	public String ATTR_DATA_SWIPE_CLOSE = "data-swipe-close";
-	public String ATTR_DATA_THEME = "data-theme";
-	public String ATTR_DATA_TRACK_THEME = "data-track-theme";
-	public String ATTR_DATA_DIVIDER_THEME = "data-divider-theme";
-	public String ATTR_DATA_TRANSITION = "data-transition";
-	public String ATTR_DATA_TYPE = "data-type";
-	public String ATTR_DATA_CORNERS = "data-corners";
-	public String ATTR_DATA_POSITION_TO = "data-position-to";
-
-	public String ATTR_DATA_MIN = "min";
-	public String ATTR_DATA_MAX = "max";
-	public String ATTR_DATA_STEP = "step";
-	public String ATTR_DATA_VALUE = "value";
-	public String ATTR_PLACEHOLDER = "placeholder";
-	public String ATTR_PATTERN = "pattern";
-	public String ATTR_DATA_DISMISSABLE = "data-dismissible";
-	public String ATTR_DATA_SHADOW = "data-shadow";
-	public String ATTR_AUTOFOCUS = "autofocus";
-	public String ATTR_REQUIRED = "required";
-	public String ATTR_MAXLENGTH = "maxlength";
-	public String ATTR_AUTOCOMPLETE = "autocomplete";
-	public String ATTR_NOVALIDATE = "novalidate";
-
-	public String TRUE = "true";
-	public String FALSE = "false";
-	public String TYPE_CHECKBOX = "checkbox";
-	public String TYPE_RADIO = "radio";
-	public String TYPE_RANGE = "range";
-	public String TYPE_TEXT = "text";
-	public String TYPE_TEXTAREA = "textarea";
-	public String TYPE_SEARCH = "search";
-	public String TYPE_PASSWORD = "password";
-	public String TYPE_FILE = "file";
-	public String TYPE_NUMBER = "number";
-	public String TYPE_URL = "url";
-	public String TYPE_COLOR = "color";
-	public String TYPE_EMAIL = "email";
-	public String TYPE_TEL = "tel";
-	public String TYPE_DATE = "date";
-	public String TYPE_MONTH = "month";
-	public String TYPE_WEEK = "week";
-	public String TYPE_TIME = "time";
-	public String TYPE_DATETIME = "datetime";
-
-	public String CLASS_CUSTOM = "custom";
-	public String CLASS_DISABLED = "ui-disabled";
-	public String CLASS_HIDDEN_ACCESSIBLE = "ui-hidden-accessible";
-	public String CLASS_HIDE_LABEL = "ui-hide-label";
-	public String CLASS_BUTTON_RIGHT = "ui-btn-right";
-	public String CLASS_BUTTON_LEFT = "ui-btn-left";
-	public String CLASS_BAR = "ui-bar";
-	public String CLASS_CONTENT = "ui-content";
-	public String CLASS_RESPONSIVE = "ui-responsive";
-	public String CLASS_TABLE_STRIPE = "table-stripe";
-
-	public String POSITION_FIXED = "fixed";
-	public String POSITION_TO_WINDOW = "window";
-	public String POSITION_TO_ORIGIN = "origin";
-
-	public String POSITION_LEFT = "left";
-	public String POSITION_RIGHT = "right";
-
-	public String DISPLAY_OVERLAY = "overlay";
-	public String DISPLAY_REVEAL = "reveal";
-	public String DISPLAY_PUSH = "push";
-
-	public String MODE_COLUMNTOGGLE = "columntoggle";
-	public String MODE_REFLOW = "reflow";
-
-	public String ROLE_BUTTON = "button";
-	public String ROLE_COLLAPSIBLE = "collapsible";
-	public String ROLE_COLLAPSIBLE_SET = "collapsible-set";
-	public String ROLE_CONTENT = "content";
-	public String ROLE_DIALOG = "dialog";
-	public String ROLE_FIELDCONTAIN = "fieldcontain";
-	public String ROLE_FOOTER = "footer";
-	public String ROLE_GROUP = "controlgroup";
-	public String ROLE_HEADER = "header";
-	public String ROLE_LISTVIEW = "listview";
-	public String ROLE_DIVIDER = "list-divider";
-	public String ROLE_NAVBAR = "navbar";
-	public String ROLE_PAGE = "page";
-	public String ROLE_POPUP = "popup";
-	public String ROLE_RANGE_SLIDER = "rangeslider";
-	public String ROLE_SLIDER = "slider";
-	public String ROLE_TABLE = "table";
-
-	public String DATA_TYPE_HORIZONTAL = "horizontal";
-
-	public String LAYOUT_HORIZONTAL = "Horizontal";
-	public String LAYOUT_VERTICAL = "Vertical";
-
-	public String TRANSITION_FADE = "fade"; //default
-	public String TRANSITION_POP = "pop";
-	public String TRANSITION_FLIP = "flip";
-	public String TRANSITION_TURN = "turn";
-	public String TRANSITION_FLOW = "flow";
-	public String TRANSITION_SLIDEFADE = "slidefade";
-	public String TRANSITION_SLIDEDOWN = "slidedown";
-	public String TRANSITION_SLIDE = "slide";
-	public String TRANSITION_SLIDEUP = "slideup";
-	public String TRANSITION_NONE = "none";
-
-	public String DATA_REL_DIALOG = "dialog";
-	public String DATA_REL_POPUP = "popup";
-	public String DATA_REL_BACK = "back";
-	public String DATA_REL_CLOSE = "close";
-	public String DATA_REL_EXTERNAL = "external";
-	public String CLOSE_LEFT = "left";
-	public String CLOSE_RIGHT = "right";
-	public String CLOSE_NONE = "none";
-
-	public String ARRAGEMENT_DEFAULT = "Default";
-	public String ARRAGEMENT_GROUPED = "Grouped";
-	public String ARRAGEMENT_NAVBAR = "Navbar";
-
-	public String ICONPOS_NOTEXT = "notext";
-
-	public String AUTOCOMPLETE_OFF = "off";
-
+public interface JQueryConstants extends JQueryHTMLConstants {
 	public String EDITOR_ID_TITLE = TAG_TITLE;
 	public String EDITOR_ID_LABEL = TAG_LABEL;
 	public String EDITOR_ID_NAME = ATTR_NAME;
