+++ plugins/org.jboss.tools.jst.web/META-INF/MANIFEST.MF
@@ -32,6 +32,7 @@ Export-Package:
  org.jboss.tools.jst.web.browser,
  org.jboss.tools.jst.web.browser.wtp,
  org.jboss.tools.jst.web.context,
+ org.jboss.tools.jst.web.html,
  org.jboss.tools.jst.web.launching.sourcelookup,
  org.jboss.tools.jst.web.launching.sourcelookup.xpl,
  org.jboss.tools.jst.web.messages.xpl,
